+++ generator/src/main/java/com/scottlogic/deg/generator/generation/DataGeneratorMonitor.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.generation;
 
-import com.scottlogic.deg.generator.generation.databags.Row;
+import com.scottlogic.deg.generator.generation.rows.Row;
 
 public interface DataGeneratorMonitor {
     void generationStarting(GenerationConfig generationConfig);
