+++ generator/src/main/java/com/scottlogic/deg/generator/generation/PartitioningRowSolver.java
@@ -5,7 +5,7 @@ import com.scottlogic.deg.generator.decisiontree.DecisionTree;
 import com.scottlogic.deg.generator.decisiontree.DecisionTreeOptimiser;
 import com.scottlogic.deg.generator.decisiontree.treepartitioning.TreePartitioner;
 import com.scottlogic.deg.generator.generation.combinationstrategies.CombinationStrategy;
-import com.scottlogic.deg.generator.generation.databags.*;
+import com.scottlogic.deg.generator.generation.rows.*;
 
 import java.util.stream.Stream;
