+++ generator/src/main/java/com/scottlogic/deg/generator/generation/combinationstrategies/CombinationStrategy.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.generation.combinationstrategies;
 
-import com.scottlogic.deg.generator.generation.databags.Row;
+import com.scottlogic.deg.generator.generation.rows.Row;
 
 import java.util.stream.Stream;
