+++ generator/src/main/java/com/scottlogic/deg/generator/generation/combinationstrategies/PinningCombinationStrategy.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.generation.combinationstrategies;
 
-import com.scottlogic.deg.generator.generation.databags.Row;
-import com.scottlogic.deg.generator.generation.databags.GeneratedObjectMerger;
+import com.scottlogic.deg.generator.generation.rows.Row;
+import com.scottlogic.deg.generator.generation.rows.GeneratedObjectMerger;
 
 import java.util.Collections;
 import java.util.Iterator;
