+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/FileOutputTarget.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.outputs.targets;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.scottlogic.deg.generator.ProfileFields;
-import com.scottlogic.deg.generator.generation.databags.Row;
+import com.scottlogic.deg.generator.generation.rows.Row;
 import com.scottlogic.deg.generator.outputs.datasetwriters.DataSetWriter;
 
 import java.io.Closeable;
