+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/OutputTarget.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.outputs.targets;
 
 import com.scottlogic.deg.generator.ProfileFields;
-import com.scottlogic.deg.generator.generation.databags.Row;
+import com.scottlogic.deg.generator.generation.rows.Row;
 
 import java.io.IOException;
 import java.util.stream.Stream;
