+++ generator/src/test/java/com/scottlogic/deg/generator/generation/rows/RowTests.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.generation.databags;
+package com.scottlogic.deg.generator.generation.rows;
 
 import com.scottlogic.deg.generator.Value;
 import com.scottlogic.deg.generator.Field;
